Updated the authentication and worked on the  schoolportal

+++ Services/FIileService.cs
@@ -1,5 +1,4 @@
 ﻿
-using StudentProject.ViewModel;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
@@ -8,8 +7,10 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
+using StudentProject.ViewModels;
+using System.Runtime.InteropServices;
 
-namespace EmployeeManagementwithdatabase1.Services
+namespace StudentProject.Services
 {
     public class FileService : IFileManagerService
     {
@@ -20,21 +21,50 @@ namespace EmployeeManagementwithdatabase1.Services
         {
             _videopath = configuration["Path:Videos"];
             this._Logger=Logger;
-
         }
-
         public string _videopath { get; }
 
-        public void DeleteImage(EditViewModel editView)
+        public void DeleteVideo(string videopath)
         {
-            try{
-            var save_path = Path.Combine(_videopath,editView.ExistingPhotoPath);
-            File.Delete(save_path);
-            }catch{
+            try
+            {
+                var save_path = Path.Combine(_videopath, videopath);
+                //if (Directory.Exists(save_path))
+                //{
+                //    File.Delete(save_path);
+                //    Directory.Delete(save_path);
+                //}
+                if (System.IO.File.Exists(videopath))
+                {
+                    System.IO.File.Delete(videopath);
+                }
+
+            }
+            catch
+            {
+                _Logger.LogInformation("Error");
+            }
+        }
+        public void DeleteVideo(EditSpellerViewModel editView)
+        {
+            try
+            {
+                var save_path = Path.Combine(_videopath, editView.ExistingVideoPath);
+                if (Directory.Exists(save_path))
+                {
+                    File.Delete(save_path);
+                    Directory.Delete(save_path);
+                }
+            }
+            catch
+            {
                 _Logger.LogInformation("Error");
             }
         }
 
+       
+
+       
         public async Task<string> SaveVideo(IFormFile video)
         {
             try
@@ -42,9 +72,11 @@ namespace EmployeeManagementwithdatabase1.Services
                 var save_path = Path.Combine(_videopath);
                 if (!Directory.Exists(_videopath))
                 {
+                   
                     Directory.CreateDirectory(save_path);
                 }
-                var filename = new Guid() + video.FileName;
+                var guid = Guid.NewGuid();
+                var filename = guid + video.FileName;
                 using (var filestream = new FileStream(Path.Combine(save_path, filename), FileMode.Create))
                 {
             
@@ -60,6 +92,16 @@ namespace EmployeeManagementwithdatabase1.Services
                 return "Error";
             }
         }
-       
+
+        public async Task<string> UpdateVideo(EditSpellerViewModel editSpeller)
+        {
+            if(editSpeller.VideoPath != null)
+            {
+                DeleteVideo(editSpeller);
+                var filename = await SaveVideo(editSpeller.VideoPath);
+                return filename;
+            }
+            return editSpeller.ExistingVideoPath;
+        }
     }
 }
